Refactor user management actions and enhance utility functions
- Introduced AdminUserActions and InvitationActions components for better handling of user and invitation management.
- Removed the UserRowActions component to streamline the codebase.
- Moved getInitials and formatDate functions to utils for reusability across components.
- Updated UserManagementTable to utilize new action components for improved clarity and functionality.
